+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 
 * Include all SDL2 needed definitions
 * Try to support games calling vk functions directly
+* When both SDL2 and SDL3 functions exist with same name
+  and different arguments, implement a trampoline function
+  that can forward arguments to the right function 
 
 ### Fixed